refactor: Move constants to the `constants.ts` file
