+++ docs/classes/MiniCSS.html
@@ -75,28 +75,28 @@
 <div class="tsd-comment tsd-typography"><p>CSS class map.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/minicss/core/blob/547ab2f/src/MiniCSS.ts#L13">src/MiniCSS.ts:13</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/minicss/core/blob/cfeea37/src/MiniCSS.ts#L13">src/MiniCSS.ts:13</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class tsd-is-private"><a id="_ids" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagPrivate">Private</code> <span>#ids</span><a href="#_ids" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">#ids<span class="tsd-signature-symbol">:</span> <a href="Node.html" class="tsd-signature-type" data-tsd-kind="Class">Node</a><span class="tsd-signature-symbol"> = ...</span></div>
 <div class="tsd-comment tsd-typography"><p>CSS id map.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/minicss/core/blob/547ab2f/src/MiniCSS.ts#L19">src/MiniCSS.ts:19</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/minicss/core/blob/cfeea37/src/MiniCSS.ts#L19">src/MiniCSS.ts:19</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class tsd-is-private"><a id="_keyframes" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagPrivate">Private</code> <span>#keyframes</span><a href="#_keyframes" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">#keyframes<span class="tsd-signature-symbol">:</span> <a href="Node.html" class="tsd-signature-type" data-tsd-kind="Class">Node</a><span class="tsd-signature-symbol"> = ...</span></div>
 <div class="tsd-comment tsd-typography"><p>CSS keyframes map.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/minicss/core/blob/547ab2f/src/MiniCSS.ts#L25">src/MiniCSS.ts:25</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/minicss/core/blob/cfeea37/src/MiniCSS.ts#L25">src/MiniCSS.ts:25</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class tsd-is-private"><a id="_variables" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagPrivate">Private</code> <span>#variables</span><a href="#_variables" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">#variables<span class="tsd-signature-symbol">:</span> <a href="Node.html" class="tsd-signature-type" data-tsd-kind="Class">Node</a><span class="tsd-signature-symbol"> = ...</span></div>
 <div class="tsd-comment tsd-typography"><p>CSS variable map.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/minicss/core/blob/547ab2f/src/MiniCSS.ts#L31">src/MiniCSS.ts:31</a></li></ul></aside></section></section>
+<li>Defined in <a href="https://github.com/minicss/core/blob/cfeea37/src/MiniCSS.ts#L31">src/MiniCSS.ts:31</a></li></ul></aside></section></section>
 <section class="tsd-panel-group tsd-member-group">
 <h2>Methods</h2>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class"><a id="class" class="tsd-anchor"></a>
