+++ docs/classes/Node.html
@@ -79,11 +79,11 @@
 <div class="tsd-comment tsd-typography"><p>Last generated name.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/minicss/core/blob/547ab2f/src/Node.ts#L12">src/Node.ts:12</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/minicss/core/blob/cfeea37/src/Node.ts#L12">src/Node.ts:12</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class tsd-is-inherited tsd-is-external"><a id="_toStringTag_" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>[to<wbr/>String<wbr/>Tag]</span><a href="#_toStringTag_" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">[to<wbr/>String<wbr/>Tag]<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div><aside class="tsd-sources">
-<p>Inherited from Map.__@toStringTag@124</p>
+<p>Inherited from Map.__@toStringTag@129</p>
 <ul>
 <li>Defined in node_modules/typescript/lib/lib.es2015.symbol.wellknown.d.ts:135</li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class tsd-is-inherited tsd-is-external"><a id="size" class="tsd-anchor"></a>
