+++ src/index.ts
@@ -1,3 +1,3 @@
 export { default, default as MiniCSS } from "./MiniCSS.js";
 export { default as Node } from "./Node.js";
-export { type MiniCSSJSONOutputI } from "./utils.js";
+export { type MiniCSSJSONOutputI } from "./constants.js";
